refactor: centralize MockLogger in internal/testutils
- Create unified MockLogger in internal/testutils
- Remove duplicate mock implementations from tests/mocks
- Update all test files to use testutils.NewMockLogger()
- Fix broken references and imports across test suite
