+++ hotkeys/manager/manager_extended_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/AshBuk/speak-to-ai/hotkeys/interfaces"
 	"github.com/AshBuk/speak-to-ai/hotkeys/mocks"
 	"github.com/AshBuk/speak-to-ai/hotkeys/utils"
+	"github.com/AshBuk/speak-to-ai/internal/testutils"
 )
 
 func TestParseHotkey(t *testing.T) {
@@ -169,7 +170,7 @@ func TestHotkeyManager_StartWithProviderFailure(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+r", "auto")
 
 	// Test when provider fails to start
-	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, newMockLogger())
+	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Replace with a mock that always fails
 	mockProvider := mocks.NewMockHotkeyProvider()
@@ -218,7 +219,7 @@ func TestSelectKeyboardProvider(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			config := adapters.NewConfigAdapter("ctrl+r", "auto")
-			manager := NewHotkeyManager(config, tt.environment, newMockLogger())
+			manager := NewHotkeyManager(config, tt.environment, testutils.NewMockLogger())
 
 			if manager == nil {
 				t.Fatal("NewHotkeyManager returned nil")
@@ -237,7 +238,7 @@ func TestSelectKeyboardProvider(t *testing.T) {
 
 func TestHotkeyManager_StopWithoutStart(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+r", "auto")
-	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, newMockLogger())
+	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Stop without start should be safe
 	manager.Stop()
@@ -249,7 +250,7 @@ func TestHotkeyManager_StopWithoutStart(t *testing.T) {
 
 func TestHotkeyManager_MultipleStartStop(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+r", "auto")
-	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, newMockLogger())
+	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 	mockProvider := mocks.NewMockHotkeyProvider()
 	manager.provider = mockProvider
 
@@ -284,7 +285,7 @@ func TestHotkeyManager_MultipleStartStop(t *testing.T) {
 
 func TestHotkeyManager_ConcurrentAccess_Extended(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+r", "auto")
-	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, newMockLogger())
+	manager := NewHotkeyManager(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 	mockProvider := mocks.NewMockHotkeyProvider()
 	manager.provider = mockProvider
