+++ hotkeys/providers/dbus_accelerator_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/AshBuk/speak-to-ai/hotkeys/adapters"
 	"github.com/AshBuk/speak-to-ai/hotkeys/interfaces"
+	"github.com/AshBuk/speak-to-ai/internal/testutils"
 )
 
 func TestConvertHotkeyToAccelerator(t *testing.T) {
@@ -64,7 +65,7 @@ func TestConvertHotkeyToAccelerator(t *testing.T) {
 func TestDbusKeyboardProvider_IsSupported(t *testing.T) {
 	config := adapters.NewConfigAdapter("altgr+comma", "auto")
 
-	provider := NewDbusKeyboardProvider(config, interfaces.EnvironmentWayland, newMockLogger())
+	provider := NewDbusKeyboardProvider(config, interfaces.EnvironmentWayland, testutils.NewMockLogger())
 
 	// Test that IsSupported doesn't panic
 	// The result depends on system environment, so we just check it doesn't crash
@@ -75,7 +76,7 @@ func TestDbusKeyboardProvider_IsSupported(t *testing.T) {
 func TestDbusProvider_RegisterHotkey(t *testing.T) {
 	config := adapters.NewConfigAdapter("altgr+comma", "auto")
 
-	provider := NewDbusKeyboardProvider(config, interfaces.EnvironmentWayland, newMockLogger())
+	provider := NewDbusKeyboardProvider(config, interfaces.EnvironmentWayland, testutils.NewMockLogger())
 
 	// Test callback registration
 	callback := func() error {
