+++ hotkeys/providers/evdev_provider_test.go
@@ -8,13 +8,14 @@ import (
 
 	"github.com/AshBuk/speak-to-ai/hotkeys/adapters"
 	"github.com/AshBuk/speak-to-ai/hotkeys/interfaces"
+	"github.com/AshBuk/speak-to-ai/internal/testutils"
 )
 
 func TestNewEvdevKeyboardProvider(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+shift+r", "auto")
 	env := interfaces.EnvironmentX11
 
-	provider := NewEvdevKeyboardProvider(config, env, newMockLogger())
+	provider := NewEvdevKeyboardProvider(config, env, testutils.NewMockLogger())
 
 	if provider == nil {
 		t.Fatal("NewEvdevKeyboardProvider returned nil")
@@ -42,7 +43,7 @@ func TestNewEvdevKeyboardProvider(t *testing.T) {
 
 func TestEvdevKeyboardProvider_IsSupported(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+shift+r", "auto")
-	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, newMockLogger())
+	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Test IsSupported - this will likely return false in test environment due to permissions
 	supported := provider.IsSupported()
@@ -58,7 +59,7 @@ func TestEvdevKeyboardProvider_IsSupported(t *testing.T) {
 
 func TestEvdevKeyboardProvider_RegisterHotkey(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+shift+r", "auto")
-	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, newMockLogger())
+	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	callbackCalled := false
 	callback := func() error {
@@ -96,7 +97,7 @@ func TestEvdevKeyboardProvider_RegisterHotkey(t *testing.T) {
 
 func TestEvdevKeyboardProvider_Start_AlreadyStarted(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+shift+r", "auto")
-	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, newMockLogger())
+	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Set isListening to true to simulate already started
 	provider.isListening = true
@@ -112,7 +113,7 @@ func TestEvdevKeyboardProvider_Start_AlreadyStarted(t *testing.T) {
 
 func TestEvdevKeyboardProvider_Stop_NotStarted(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+shift+r", "auto")
-	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, newMockLogger())
+	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Stop should not panic even if not started
 	provider.Stop()
@@ -125,7 +126,7 @@ func TestEvdevKeyboardProvider_Stop_NotStarted(t *testing.T) {
 
 func TestEvdevKeyboardProvider_Stop_WhenStarted(t *testing.T) {
 	config := adapters.NewConfigAdapter("ctrl+shift+r", "auto")
-	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, newMockLogger())
+	provider := NewEvdevKeyboardProvider(config, interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Simulate started state
 	provider.isListening = true
@@ -250,7 +251,7 @@ func TestHasKeyEvents(t *testing.T) {
 
 func TestModifierStateTracking(t *testing.T) {
 	// Test that modifier state tracking logic works
-	provider := NewEvdevKeyboardProvider(adapters.NewConfigAdapter("ctrl+shift+r", "auto"), interfaces.EnvironmentX11, newMockLogger())
+	provider := NewEvdevKeyboardProvider(adapters.NewConfigAdapter("ctrl+shift+r", "auto"), interfaces.EnvironmentX11, testutils.NewMockLogger())
 
 	// Simulate modifier key press
 	provider.modifierState["leftctrl"] = true
