+++ tests/integration/platform_integration_test.go
@@ -15,6 +15,7 @@ import (
 	hotkeyinterfaces "github.com/AshBuk/speak-to-ai/hotkeys/interfaces"
 	"github.com/AshBuk/speak-to-ai/hotkeys/providers"
 	"github.com/AshBuk/speak-to-ai/internal/platform"
+	"github.com/AshBuk/speak-to-ai/internal/testutils"
 	outputfactory "github.com/AshBuk/speak-to-ai/output/factory"
 )
 
@@ -45,7 +46,7 @@ func TestWaylandCompatibility(t *testing.T) {
 	t.Run("wayland_hotkeys", func(t *testing.T) {
 		// Test DBus hotkey provider
 		hotkeyConfig := adapters.NewConfigAdapter("altgr+comma", "auto")
-		provider := providers.NewDbusKeyboardProvider(hotkeyConfig, hotkeyinterfaces.EnvironmentWayland, newMockLogger())
+		provider := providers.NewDbusKeyboardProvider(hotkeyConfig, hotkeyinterfaces.EnvironmentWayland, testutils.NewMockLogger())
 
 		if provider.IsSupported() {
 			t.Log("DBus GlobalShortcuts portal is available")
@@ -91,7 +92,7 @@ func TestX11Compatibility(t *testing.T) {
 	t.Run("x11_hotkeys", func(t *testing.T) {
 		// Test evdev hotkey provider
 		hotkeyConfig := adapters.NewConfigAdapter("altgr+comma", "auto")
-		provider := providers.NewEvdevKeyboardProvider(hotkeyConfig, hotkeyinterfaces.EnvironmentX11, newMockLogger())
+		provider := providers.NewEvdevKeyboardProvider(hotkeyConfig, hotkeyinterfaces.EnvironmentX11, testutils.NewMockLogger())
 
 		if provider.IsSupported() {
 			t.Log("Evdev input devices are available")
